+++ CHANGELOG.md
@@ -46,6 +46,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Dependencies versions bumped ([#277](https://github.com/EpiLink/EpiLink/pull/277), [#280](https://github.com/EpiLink/EpiLink/pull/280) and [#287](https://github.com/EpiLink/EpiLink/pull/287))
 * Updated the Discord logo and color ([#283](https://github.com/EpiLink/EpiLink/pull/283))
 * Migrated front-end to Vue 3 ([#316](https://github.com/EpiLink/EpiLink/pull/316))
+* Removed the italics on the front-end that were not used properly ([#325](https://github.com/EpiLink/EpiLink/pull/325))
+* The "About EpiLink" page is now opt-in and can be enabled in the configuration ([#325](https://github.com/EpiLink/EpiLink/pull/325))
 * The following are worth mentioning, though they are internal changes that don't impact the actual feature set. ([#280](https://github.com/EpiLink/EpiLink/pull/280))
   * The CI is now more thorough in order to improve our release quality.
   * The front-end is now bundled as a separate JAR that is *then* integrated as a regular library in the back-end. This significantly simplifies the `withFrontend` variant build.